[PR #5140 @dilijev] Add ChakraEngine for GlobalObject lifetime management
ChakraEngine behavior in jshost is guarded by a host flag, and in Chakra is guarded on the presence of (chakraEngine != nullptr) on ScriptEngine and JavascriptLibrary.

When ChakraEngine behavior is not enabled, behavior should be unchanged from before, and passes all unit tests.

With ChakraEngine behavior enabled, there are still some assertion hits and leaks. I'm going to keep working on these issues. I'm going to add some tests which use the -UseChakraEngineAPI flag

[Core Commit Message]

[MERGE #5140 @dilijev] Add ChakraEngine for GlobalObject lifetime management

Merge pull request #5140 from dilijev:ChakraEngine