[PR #5150 @thomasmo] Enable Deferred Parsing for Background Parse
In the Full part of the change, jshost will now run tests with BGParse when there is an addition host flag specified, ExecuteWithBgParse. This flag is enabled on a few Core and Full tests to ensure that this code is exercised regularly.
Further, ScriptEngine is updated to fall back to synchronous parsing when there is a failure with getting results from BgParse to ensure self-hosting experiences.

[Core Commit Message]

[MERGE #5150 @thomasmo] Enable Deferred Parsing for Background Parse

Merge pull request #5150 from thomasmo:bgparse_defer

This change turns on Deferred Parsing in BGParseManager, which will reduce the amount of time needed to parse in the background. Appropriate flags are added in calls inside BGParseManager and BGParseWorkItem.
Also, in order to serialize defer parsed bytecode, ScriptContext::CompileUTF8Core refactors the Parser object outside of the call because the serializer needs state from the parser to complete.