[PR #5119 @mike-kaufman] Adding capture & transmission of Recycler telemetry
Updated Recycler & associated classes to capture & transmit GC-related telemetry data.  This includes:

  - PageAllocator sizes
  - Total Bytes used by the recycler (across all HeapBuckets)
  - Amount of time GC blocks the UI thread
  - Number of pages de-committed for each allocator.
  -  Microsecond time values for how long we spend computing the telemetry statistics.

The actual data we collect will evolve as we start digging into it

We define a "GC Pass" as the period from when the Recycler's CollectionState  transitions from `Not Collecting` and then transitions back to `Not Collecting` . We compute most stats based on this "Pass".  We currently attempt to transmit data when we have 16 passes accumulated.  Again, this will be tweaked as dig into specific telemetry data flowing back indicating frequency of GC passes & frequency of transmissions.

Capturing MemGC stats is out-of-scope at this time.

There are still some TODOs pending based on early feedback, namely:
  - [ ] Run benchmarks and get a feel for impact on those.
  - [x] Fine-tune the assertions about which thread we're running on
  - [x] include indicators if Concurrent GC is enabled.

[Core Commit Message]

[MERGE #5119 @mike-kaufman] Adding capture & transmission of Recycler telemetry

Merge pull request #5119 from mike-kaufman:build/mkaufman/gc-telemetry

Updated Recycler & associated classes to capture & transmit GC-related telemetry data.  This includes:

  - PageAllocator sizes
  - Total Bytes used by the recycler (across all HeapBuckets)
  - Amount of time GC blocks the UI thread
  - Number of pages de-committed for each allocator.
  -  Microsecond time values for how long we spend computing the telemetry statistics.

The actual data we collect will evolve as we start digging into it

We define a "GC Pass" as the period from when the Recycler's CollectionState  transitions from `Not Collecting` and then transitions back to `Not Collecting` . We compute most stats based on this "Pass".  We currently attempt to transmit data when we have 16 passes accumulated.  Again, this will be tweaked as dig into specific telemetry data flowing back indicating frequency of GC passes & frequency of transmissions.

Capturing MemGC stats is out-of-scope at this time.

There are still some TODOs pending based on early feedback, namely:
  - [ ] Run benchmarks and get a feel for impact on those.
  - [x] Fine-tune the assertions about which thread we're running on
  - [x] include indicators if Concurrent GC is enabled.