[MERGE #5147 @rhuanjl] Fix errors in construction of bound Functions
Merge pull request #5147 from rhuanjl:boundProxy

Revisiting boundProxies found what I hope are the right fixes this time.

This PR addresses the following issues:
1. Constructing bound Proxy would call construct trap twice or if there wasn't a construct trap targetFunction would be called twice,
2. Reflect.construct in construct trap of bound Proxy would attempt to construct "this" and error as it's not a constructor (issue 4918)
3. Constructing a bound function via Reflect with a newTarget would ignore the newTarget.

Added test cases for these points and some related patterns that do work but weren't being tested and  which some other ways of fixing points 1-3 would break.

Also corrected the baseline for proxytest9.js which was testing the incorrect behaviour of construct traps for bound proxies being called twice.

The interaction of bound functions and proxies is complex enough that I'm afraid I'm confident that there'll be one or more bugs left after this, but this should be a reasonable improvement from the current state.

@sethbrenith I borrowed your test case from 4918 as the starting point for the cases here, I hope that's ok. Note I split it into two as it highlighted 2 different issues.

Fixes: #4918
Fixes: #5151