[MERGE #5175 @akroshg] Constructing object on chained proxy
Merge pull request #5175 from akroshg:proxy

In the functiontrap we don't create an object if the target itself is proxy. However as we provide NewTarget flag, the nested functiontrap will try to take the args[0] which later
leads to Assert as it was null.
Fixed that by constructing an object if the args[0] is null even if the NewTarget flag is passed.