[MERGE #5181 @akroshg] Incorrect object shorthand error in lambda
Merge pull request #5181 from akroshg:shorthand

The object definition at lambda param should be treated as object pattern. However we had deduced that we are on right side of an expression. Which is clearly not the case.
Fixed that, by let the statement boundary determines that there is any shorthand error deferred and then throw error.