[MERGE #5207 @MikeHolman] add NaN check for simple Sets
Merge pull request #5207 from MikeHolman:setnan

It is possible for sign bit to be set on NaNs (e.g. when they are the result of some math operations), so canonicalize them when adding to simple set in order to avoid duplicating entries.

We can do this because the sign bit on NaN is not observable to JS code.

OS: 17498790