Merged PR 124697: OS#17625024 - Fix ProfileCacheTests native unit tests
OS#17625024 - Fix ProfileCacheTests native unit tests

We added a second flavor for this unit test but it reads and writes to files with fixed names so the two tests stomp on each other. Fixed by having the filenames be generated based on a parameter passed to the native tests dll.

Fixes:
https://microsoft.visualstudio.com/_workitems/edit/17625024