Fixing a race condition in ScriptSite::Close
ScriptSite::Close would not mark itself as closed until toward
the end of its cleanup. However, early on it closes the projectionContext
which frees that memory, and then later cleanup can cause the recycler
to finish cleaning up in-thread, which can cause finalizers to run and
try to access the freed memory. By marking the ScriptSite as closed,
that finalizer will no longer assume that the projectionContext is valid
and so we avoid the use-after-free.