[PR #5332 @paolosevMSFT] OS#17885560 - Fix perf regression introduced with a bad fix for OS#17417473 [Update submodule only]
The fix for bug 17417473 was incorrect: in GlobOpt::OptIsInvariant() we should use loop->landingPad->globOptData.liveVarSyms, not loop->landingPad->globOptData.liveFields to make sure a symbol is not hoisted if not alive in landingPad.
This also introduced a perf regression; in test mandreel.js a large number of "CheckFixedFld" instructions were not hoisted outside loops.

[Core Commit Message]

[MERGE #5332 @paolosevMSFT] OS#17885560 - Fix perf regression introduced with a bad fix for OS#17417473

Merge pull request #5332 from paolosevMSFT:17885560

The fix for bug 17417473 was incorrect: in GlobOpt::OptIsInvariant() we should use loop->landingPad->globOptData.liveVarSyms, not loop->landingPad->globOptData.liveFields to make sure a symbol is not hoisted if not alive in landingPad.
This also introduced a perf regression; in test mandreel.js a large number of "CheckFixedFld" instructions were not hoisted outside loops.