[MERGE #5337 @sigatrev] OS#17666127: (ARM64) fix bailout problem from indir offset legalization
Merge pull request #5337 from sigatrev:arm64lea

In ARM64 LEA is canonicalized into an ADD, and both operands of ADDs must be the same size, either 64 or 32. For LEAs with 64 bit bases and 32 bit offsets, this means converting the 32 bit offset to a 64 bit value with sign extention as needed. Calling UseWithNewType on an operand with the original 32 bit symbol resulted in the symbol's type changing as well, so the register allocator would not know that the symbol was supposed to be a tagged int, and it would not be retagged on bailout.