[MERGE #5339 @MikeHolman] avoid calling GetCallSiteInfo from background
Merge pull request #5339 from MikeHolman:calleeinfobug

GetCallSiteInfo isn't thread safe, so avoid calling it from the background.

It was only ever called in the background from FunctionBody::HasNonBuiltInCallee, which only needs some of the info. So I've made a new method that only reads what we need and put it/updates to CallSiteInfo behind a critical section.

OS: 17904582