[MERGE #5328 @mike-kaufman] fixing issue where "uncaught" exceptions in promises wouldn't notify debugger
Merge pull request #5328 from mike-kaufman:build/mkaufman/support-uncaught-exception-handler-in-promises

If an exception was raised inside a promise and the promise didn't have any rejection handlers, we wouldn't notify the debugger that an "unhandled" exception occurred.  Fixed this up and added some simple tests for it.

This addresses the common cases for promises, but doesn't yet address async/await constructs.  Will leave #4630 open for that.