[MERGE #5399 @pleath] OS#16526223: Keep writes to 'prototype' from going through inline cache so that we get the necessary runtime behavior
Merge pull request #5399 from pleath:16526223

Writes to the prototype property of JavascriptFunction must go through the runtime so that the constructor cache, etc., can be invalidated. Enumeration provided a way to record a write to 'prototype' in an inline cache so that the runtime could be skipped. One solution would be to make 'prototype' an accessor property on JavascriptFunction, but this would probably grow function objects and slow down loads of the property. Instead, let PropertyRecord indicate that we should disable caching of writes to the property and disable store field cache in PropertyValueInfo.