[MERGE #5403 @sanketj] Scan interior pointers from GC stack roots when hosted by a native GC host
Merge pull request #5403 from sanketj:user/sajos/scan_interior_pointers_from_stack_for_native_gc_host

Currently, we skip scanning interior pointers from GC stack roots if we are not running script (or if we are in mem protect mode). When host objects are also placed on the Chakra heap for GC purposes, we must scan interior pointers from the stack even when not running script. To scope this fix to just native GC host scenarios, and to avoid regressing JSRT scenarios, we are introducing a bit on the recycler that will be set for native GC hosts. If this bit is set, we will always scan interior pointers from the stack, and if it is not set, we will continue doing what we do currently.

In order to set this bit, an API is being exposed on the JS thread service that is published and can be called by the host.