[MERGE #5457 @pleath] OS#18076647: Invalidate prototype inline caches for properties of objects added to the prototype chain
Merge pull request #5457 from pleath:18076647

Currently we invalidate proto inline caches for properties in the old prototype chain but not in the new one. We need to invalidate properties in the new prototype chain as well to make sure that missing property caches are cleared as necessary. Since the old and new prototype chains probably overlap, do some tracking to avoid duplicated effort.