[MERGE #5446 @MSLaguana] Update function toString
Merge pull request #5446 from MSLaguana:updateFunctionToString

According to https://tc39.github.io/Function-prototype-toString-revision/
the result of toString on a function should be the source text which matches
the grammar describing the kind of function. In practice this essentially
means that toString should just be whatever the source code defining the function
is.

To support this, we need to slightly adjust the extent of a function in
the case of member list declarations to ensure that it captures the start
of the function, including any `async` or computed function name, etc.

Fixes #2914