[MERGE #5471 @aneeshdk] Fixing an issue with super restriction state in undodefer case
Merge pull request #5471 from aneeshdk:users/aneeshd/SuperUndoDeferIssue

We currently store the status of whether super call is allowed or super property is allowed in the Parser. Because of this when we do an undodefer of a member function that state information gets lost. This change is to put that information in the function node itself and use it from there while parsing super references.