[MERGE #5484 @VSadov] Fix a possible loss of property in the TypePath
Merge pull request #5484 from VSadov:fix17745531

In a rare case TinyDictionary could lose the last value added.
We use the sign bit to tag last values in bucket chains. We also use `FF` for uninitialized/empty buckets.
In extreme rare cases 127th pid to insert may hash into a a bucket still unused by the previous 126 values.
Once 127 is tagged it becomes `FF` which will look like an empty bucket - the 127 value would be lost.

The fix is simple - detect the case as described and make 127 to chain to `FF` marker. In such case it will not be the last in the chain and will not be tagged.

Fixes: OS:#17745531