[PR #5488 @VSadov] Extra OOM crash buckets.
It could be useful for the analysis purposes to know whether the state at the time of the crash was visible or not visible, if known.

[Core Commit Message]

[MERGE #5488 @VSadov] Extra OOM crash buckets.

Merge pull request #5488 from VSadov:VisibilityBuckets

It could be useful for the analysis purposes to know whether the state at the time of the crash was visible or not visible, if known.