[MERGE #5683 @VSadov] Do not lose property string in ToPropertyKey
Merge pull request #5683 from VSadov:propString

Looks like the intent of the method is to set propString to a property string if the argument happen to be one. It does not do that now. Possibly due to a typo.

In particular, methods such as HasOwnProperty check if propString is returned and would use its inline caches, but since the property string is not returned, the optimization cannot work.