[MERGE #5669 @dilijev] Make some tweaks to RegExp debug switches.
Merge pull request #5669 from dilijev:re-debug-switches

This change only affects debug output displayed under `-RegexDebug`.

Prior to this change, `-RegexDebug` has 3 "chapters" of output, all enabled with no option to disable: AST, Annotated AST, and (Bytecode) Program.

This change adds flags `-RegexDebugAST[-]` and `-RegexDebugAnnotatedAST[-]`  to disable the AST output chapters and bring focus to the resulting bytecode program, which is the most relevant for seeing how effective the regex bytecode gen was for a given case. The AST output can be noisy and distracts from this goal.

This change leaves the AST output chapters on-by-default to match existing behavior.

The AST may be interesting information when working on the Regex engine even if not doing work on optimizations. In that case the on-by-default behavior for `-RegexDebug` should probably be: `-RegexDebugAST -RegexDebugAnnotatedAST-`